scripts/tools/requirements.txt
PyQt5==5.15.9
pillow==10.2.0
numpy==1.26.3
sounddevice==0.4.6
soundfile==0.12.1